telegram_bot/requirements.txt
pyTelegramBotAPI==4.26.0
PyYAML==6.0.2
